Argument to slaw_alloc() should be called "octlen", not "quadlen".
This is an argument name in a function prototype declaration, so it is
effectively a comment.  (i. e. this change doesn't affect the program
at all.)

The argument name in the function definition in slaw.c was already
(correctly) "octlen".

(The name "quadlen" is a leftover from the slaw version 1 encoding,
which was organized into groups of four bytes called "quads".  The
current encoding, slaw version 2, is organized into groups of eight
bytes called "octs".)